issue-34474 修改域空间发布保存相关应用配置

+++ db/asset6.5-delta-schema.sql
@@ -4,7 +4,7 @@
 -- AUTHOR: 杨崇鑫  20180724
 -- REMARK: 物业缴费V6.5所需新增的字段
 ALTER TABLE `eh_asset_module_app_mappings` ADD COLUMN `contract_originId` BIGINT(20) COMMENT '合同管理应用的originId';
-ALTER TABLE `eh_asset_module_app_mappings` ADD COLUMN `contract_changeFlag` BIGINT(20) COMMENT '是否走合同变更，1、0';
+ALTER TABLE `eh_asset_module_app_mappings` ADD COLUMN `contract_changeFlag` TINYINT COMMENT '是否走合同变更，1、0';
 
 -- --------------------- SECTION END ---------------------------------------------------------
